# packages/public-api/openapi/public-api/public-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Voyant Operator API",
    "version": "0.0.0",
    "description": "Generated from the composed operator app. Do not edit by hand."
  },
  "servers": [
    {
      "url": "/",
      "description": "This deployment (same origin)"
    }
  ],
...
  "paths": {
    "/v1/public/products/{productId}/offers": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "productId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "departureId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 2
            },
            "required": false,
            "name": "locale",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Promotional offers applicable to a product (and optional departure)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": ["string", "null"]
                          },
                          "description": {
                            "type": ["string", "null"]
                          },
                          "discountType": {
                            "type": "string",
                            "enum": ["percentage", "fixed_amount"]
                          },
                          "discountValue": {
                            "type": "string"
                          },
                          "currency": {
                            "type": ["string", "null"]
                          },
                          "applicableProductIds": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "applicableDepartureIds": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "validFrom": {
                            "type": ["string", "null"]
                          },
                          "validTo": {
                            "type": ["string", "null"]
                          },
                          "minTravelers": {
                            "type": ["integer", "null"]
                          },
                          "imageMobileUrl": {
                            "type": ["string", "null"]
                          },
                          "imageDesktopUrl": {
                            "type": ["string", "null"]
                          },
                          "stackable": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "slug",
                          "description",
                          "discountType",
                          "discountValue",
                          "currency",
                          "applicableProductIds",
                          "applicableDepartureIds",
                          "validFrom",
                          "validTo",
                          "minTravelers",
                          "imageMobileUrl",
                          "imageDesktopUrl",
                          "stackable",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          }
        },
        "operationId": "getPublicProductsByProductIdOffers",
        "summary": "GET /v1/public/products/{productId}/offers",
        "tags": ["storefront"],
        "x-voyant-module": "storefront",
        "x-voyant-surface": "public-api",
        "x-voyant-api-id": "@voyant-travel/public-api#api.public",
        "x-voyant-unit-id": "@voyant-travel/public-api",
        "x-voyant-package-name": "@voyant-travel/public-api"
      }
    },
...
  },
...
}